+++ 05_multi_frmwk/pytch/data/loader_factory.py
@@ -1,42 +1,15 @@
-import os.path as op
 import torch
 import numpy as np
 import torch.utils.data as td
 import torchvision as tv
-import cv2
 
-import neutr.data.readers.kitti_reader as kitti
 import neutr.data.preprocess as pp
+import pytch.data.datasets as pds
 import config as cfg
-
-
-class Kitti2dDetectDataset:
-    def __init__(self, dset_cfg, split, transform):
-        self.transform = transform
-        drive_path = op.join(dset_cfg.PATH, 'training', 'image_2')
-        self.data_reader = kitti.KittiReader(drive_path, split, dset_cfg)
-
-    def __len__(self):
-        return len(self.data_reader)
-
-    def __getitem__(self, index):
-        if torch.is_tensor(index):
-            index = index.tolist()
-
-        example = dict()
-        example["image"] = self.data_reader.get_image(index)
-        bbox = self.data_reader.get_bboxes(index)
-        category = self.data_reader.get_categories(index)
-        example["inst"] = np.concatenate([bbox, np.ones_like(category), category], axis=-1, dtype=np.float32)
-
-        if self.transform:
-            example = self.transform(example)
-        return example
 
 
 class ToTensor:
     """Convert ndarrays in sample to Tensors."""
-
     def __call__(self, example):
         for key, val in example.items():
             if isinstance(val, np.ndarray):
@@ -45,6 +18,7 @@ class ToTensor:
         # torch image: C x H x W
         example['image'] = example['image'].permute(2, 0, 1)
         return example
+
 
 def make_dataloader(dset_cfg, split, batch_size=4, shuffle=False, num_workers=0):
     transform = get_transform(dset_cfg)
@@ -69,13 +43,14 @@ def get_transform(dset_cfg):
 
 def get_dataset(dset_cfg, split, transform):
     if dset_cfg.NAME == 'kitti':
-        dataset = Kitti2dDetectDataset(dset_cfg, split, transform)
+        dataset = pds.Kitti2dDetectDataset(dset_cfg, split, transform)
     else:
         raise ValueError(f"No dataset named {dset_cfg.NAME} is prepared")
     return dataset
 
 
 # ==========
+import cv2
 import neutr.utils.util_function as nuf
